pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["blade", "blade.container"]

[project]
name = "blade"
dynamic = [
  "version",
]
description = "BLADE: Brainstorming LLMs as Algorithm Designer through Evolution for Online Subset Selection"
readme = "README.md"
requires-python = ">=3.9"

classifiers = [
  "Intended Audience :: Science/Research",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "absl-py==2.0.0",
  "click==8.1",
  "cloudpickle==3.0.0",
  "openai==1.65.4",
  "python-dotenv==1.0.0",
  "scipy==1.11.4",
  "numpy==1.26.2",
]

[project.scripts]
funsearch = "blade.__main__:main"
